chore: split backup page into generate and backup pages
