+++ src/app/components/Tips/index.tsx
@@ -40,7 +40,7 @@ export default function Tips() {
           border: "border-purple-500",
           arrow: "text-purple-500",
           backgroundIcon: <MnemonicTipCardIcon />,
-          link: `/accounts/${accountId}/secret-key/backup`,
+          link: `/accounts/${accountId}/secret-key/generate`,
         },
       } as const),
     [accountId]
